requirements.txt
yfinance>=0.2.36
duckdb>=0.10.0
pandas-ta>=0.3.14b1
pandas>=2.0.0
backtesting>=0.6.0
quantstats>=0.0.62
plotly>=5.0.0
matplotlib>=3.7.0
anthropic>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
fastapi>=0.100.0
uvicorn>=0.20.0
pytest>=7.0.0
